Refactor validator specs
